requirements.txt
Django>=1.4.2
djangorestframework>=2.0


